Split amp-date-countdown component files
